third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/unloading-documents/001-expected.txt

This is a testharness.js-based test.
[FAIL] document.open in unload
  assert_equals: expected "0123456789" but got "01238"
Harness: the test ran to completion.
